+++ joern-cli/frontends/rubysrc2cpg/src/test/scala/io/joern/rubysrc2cpg/querying/AssignmentTests.scala
@@ -1,0 +1,61 @@
+package io.joern.rubysrc2cpg.querying
+
+import io.joern.rubysrc2cpg.testfixtures.RubyCode2CpgFixture
+import io.shiftleft.codepropertygraph.generated.Operators
+import io.shiftleft.semanticcpg.language._
+
+class AssignmentTests extends RubyCode2CpgFixture {
+
+  "CPG for code with method identifiers and literals in simple assignments" should {
+    val cpg = code("""
+        |# call instance methods
+        |a = 1
+        |b = 2
+        |a = 3
+        |b = 4
+        |c = a*b
+        |puts "Multiplication is : #{c}"
+        |""".stripMargin)
+
+    "recognise all identifier nodes" in {
+      cpg.identifier.name("a").l.size shouldBe 3
+      cpg.identifier.name("b").l.size shouldBe 3
+      cpg.identifier.name("c").l.size shouldBe 2
+    }
+
+    "recognise all literal nodes" in {
+      cpg.literal.code("1").l.size shouldBe 1
+      cpg.literal.code("2").l.size shouldBe 1
+      cpg.literal.code("3").l.size shouldBe 1
+      cpg.literal.code("4").l.size shouldBe 1
+    }
+  }
+
+  "CPG for code with multiple assignments" should {
+    val cpg = code("""
+        |a, b, c = [1, 2, 3]
+        |a, b, c = b, c, a
+        |str1, str2 = ["hello", "world"]
+        |""".stripMargin)
+
+    "recognise all identifier nodes" in {
+      cpg.identifier.name("a").l.size shouldBe 3
+      cpg.identifier.name("b").l.size shouldBe 3
+      cpg.identifier.name("c").l.size shouldBe 3
+      cpg.identifier.name("str1").l.size shouldBe 1
+      cpg.identifier.name("str2").l.size shouldBe 1
+    }
+
+    "recognise all literal nodes" in {
+      cpg.literal.code("1").l.size shouldBe 1
+      cpg.literal.code("2").l.size shouldBe 1
+      cpg.literal.code("3").l.size shouldBe 1
+      cpg.literal.code("\"hello\"").l.size shouldBe 1
+      cpg.literal.code("\"world\"").l.size shouldBe 1
+    }
+
+    "recognise all call nodes" in {
+      cpg.call.name(Operators.assignment).l.size shouldBe 3
+    }
+  }
+}
