+++ joern-cli/frontends/rubysrc2cpg/src/test/scala/io/joern/rubysrc2cpg/querying/MiscTests.scala
@@ -6,71 +6,6 @@ import io.shiftleft.semanticcpg.language._
 
 class MiscTests extends RubyCode2CpgFixture {
 
-  "CPG for code with method identifiers and literals in simple assignments" should {
-    val cpg = code("""
-        |# call instance methods
-        |a = 1
-        |b = 2
-        |a = 3
-        |b = 4
-        |c = a*b
-        |puts "Multiplication is : #{c}"
-        |""".stripMargin)
-
-    "recognise all identifier nodes" in {
-      cpg.identifier.name("a").l.size shouldBe 3
-      cpg.identifier.name("b").l.size shouldBe 3
-      cpg.identifier.name("c").l.size shouldBe 2
-    }
-
-    "recognise all literal nodes" in {
-      cpg.literal.code("1").l.size shouldBe 1
-      cpg.literal.code("2").l.size shouldBe 1
-      cpg.literal.code("3").l.size shouldBe 1
-      cpg.literal.code("4").l.size shouldBe 1
-    }
-  }
-
-  "CPG for code with class methods, members and locals in methods" should {
-    val cpg = code("""
-        |class Person
-        |  attr_accessor :name, :age
-        |
-        |  def initialize(name, age)
-        |    @name = name
-        |    @age = age
-        |  end
-        |
-        |  def greet
-        |    puts "Hello, my name is #{@name} and I am #{@age} years old."
-        |  end
-        |
-        |  def have_birthday
-        |    @age += 1
-        |    puts "Happy birthday! You are now #{@age} years old."
-        |  end
-        |end
-        |
-        |p = Person. new
-        |p.greet
-        |""".stripMargin)
-
-    "recognise all identifier and call nodes" in {
-      cpg.identifier.name("name").l.size shouldBe 1
-      cpg.identifier.name("age").l.size shouldBe 1
-      cpg.identifier.name("@name").l.size shouldBe 2
-      cpg.identifier.name("@age").l.size shouldBe 4
-      cpg.call.name("greet").size shouldBe 1
-      cpg.method.name("initialize").size shouldBe 1
-      cpg.method
-        .name("greet")
-        .size shouldBe 1
-      cpg.call.name("puts").size shouldBe 2
-      cpg.method.name("have_birthday").size shouldBe 1
-      cpg.identifier.size shouldBe 11
-    }
-  }
-
   "CPG for code with BEGIN and END blocks" should {
     val cpg = code("""
           |#!/usr/bin/env ruby
@@ -123,48 +58,6 @@ class MiscTests extends RubyCode2CpgFixture {
     }
   }
 
-  "CPG for code with square brackets as methods" should {
-    val cpg = code("""
-          |class MyClass < MyBaseClass
-          |  def initialize
-          |    @my_hash = {}
-          |  end
-          |
-          |  def [](key)
-          |    @my_hash[key.to_s]
-          |  end
-          |
-          |  def []=(key, value)
-          |    @my_hash[key.to_s] = value
-          |  end
-          |end
-          |
-          |my_object = MyClass.new
-          |
-          |""".stripMargin)
-
-    "recognise all identifier and call nodes" in {
-      cpg.method.name("\\[]").size shouldBe 1
-      cpg.method.name("\\[]=").size shouldBe 1
-      cpg.call.name(Operators.assignment).size shouldBe 3
-      cpg.method.name("initialize").size shouldBe 1
-      cpg.call.name("to_s").size shouldBe 2
-      cpg.call.name("new").size shouldBe 1
-      cpg.call.size shouldBe 9
-      cpg.identifier.name("@my_hash").size shouldBe 3
-      cpg.identifier.name("key").size shouldBe 2
-      cpg.identifier.name("value").size shouldBe 1
-      cpg.identifier.name("my_object").size shouldBe 1
-      /*
-       * FIXME
-       *  def []=(key, value) gets parsed incorrectly with parser error "no viable alternative at input 'def []=(key, value)'"
-       *  This needs a fix in the parser and update to this UT after the fix
-       * FIXME
-       *  MyClass is identified as a variableIdentifier and so an identifier. This needs to be fixed
-       */
-    }
-  }
-
   "CPG for code with defined? keyword" should {
     val cpg = code("""
         |radius = 2
@@ -204,59 +97,6 @@ class MiscTests extends RubyCode2CpgFixture {
       cpg.literal.code("Result 2: ").l.size shouldBe 1
       cpg.literal.code("Result 3: ").l.size shouldBe 1
       cpg.literal.code("Result 4: ").l.size shouldBe 1
-    }
-  }
-
-  "CPG for code with multiple assignments" should {
-    val cpg = code("""
-        |a, b, c = [1, 2, 3]
-        |a, b, c = b, c, a
-        |str1, str2 = ["hello", "world"]
-        |""".stripMargin)
-
-    "recognise all identifier nodes" in {
-      cpg.identifier.name("a").l.size shouldBe 3
-      cpg.identifier.name("b").l.size shouldBe 3
-      cpg.identifier.name("c").l.size shouldBe 3
-      cpg.identifier.name("str1").l.size shouldBe 1
-      cpg.identifier.name("str2").l.size shouldBe 1
-    }
-
-    "recognise all literal nodes" in {
-      cpg.literal.code("1").l.size shouldBe 1
-      cpg.literal.code("2").l.size shouldBe 1
-      cpg.literal.code("3").l.size shouldBe 1
-      cpg.literal.code("\"hello\"").l.size shouldBe 1
-      cpg.literal.code("\"world\"").l.size shouldBe 1
-    }
-
-    "recognise all call nodes" in {
-      cpg.call.name(Operators.assignment).l.size shouldBe 3
-    }
-  }
-
-  "CPG for code with modules" should {
-    val cpg = code("""
-        |module Module1
-        |   def method1_1
-        |   end
-        |   def method1_2
-        |   end
-        |end
-        |
-        |module Module2
-        |   def method2_1
-        |   end
-        |   def method2_2
-        |   end
-        |end
-        |""".stripMargin)
-
-    "recognise all method nodes" in {
-      cpg.method.name("method1_1").l.size shouldBe 1
-      cpg.method.name("method1_2").l.size shouldBe 1
-      cpg.method.name("method2_1").l.size shouldBe 1
-      cpg.method.name("method2_2").l.size shouldBe 1
     }
   }
 
@@ -329,41 +169,6 @@ class MiscTests extends RubyCode2CpgFixture {
         .name("some_method")
         .l
         .size shouldBe 1
-    }
-  }
-
-  "CPG for code with private/protected/public" should {
-    val cpg = code("""
-        |class SomeClass
-        |  private
-        |  def method1
-        |  end
-        |
-        |  protected
-        |  def method2
-        |  end
-        |
-        |  public
-        |  def method3
-        |  end
-        |end
-        |
-        |""".stripMargin)
-
-    "recognise all method nodes" in {
-      cpg.method
-        .name("method1")
-        .l
-        .size shouldBe 1
-      cpg.method
-        .name("method1")
-        .l
-        .size shouldBe 1
-      cpg.method
-        .name("method3")
-        .l
-        .size shouldBe 1
-
     }
   }
